requirements.txt
Django==1.10.3
djangorestframework==3.5.3
gunicorn==19.6.0
oauthlib==2.0.1
psycopg2==2.6.2
PyJWT==1.4.2
python-openid==2.2.5
python-social-auth==0.2.21
requests==2.12.1
requests-oauthlib==0.7.0
six==1.10.0
